requirement.txt
uvicorn == 0.20.0
fastapi ==0.93.0
Jinja2 == 3.1.2
python-multipart == 0.0.6
pymongo == 4.3.3
PyJWT == 2.6.0
passlib == 1.7.4
python-dotenv == 1.0.0
pydantic == 1.10.6
bcrypt == 4.0.1